Removed unnecessary prefix from list of excludes to work and update doc
In order to make it work in non UNIX-like shells (e.g. PowerShell), we'd
have to declare those excludes including path separators in a different
way, which is probably not worth the extra noise as this is a helper
script for manual use, and not part of the automated build process.